run sphinx to update doc
