+++ docs/api_explain.rst
@@ -1,0 +1,30 @@
+Explain
+-------
+
+.. rubric:: Sensitivity analysis
+
+Perform sensitivity analysis to identify feature relevances
+
+.. currentmodule:: mlcolvar.explain.sensitivity
+
+.. autosummary::
+   :toctree: autosummary
+   :template: custom-class-template.rst
+
+   sensitivity_analysis 
+   plot_sensitivity
+
+.. rubric:: Sparse linear model
+
+Use sparse models to approximate classification/regression tasks
+
+.. currentmodule:: mlcolvar.explain.lasso
+
+.. autosummary::
+   :toctree: autosummary
+   :template: custom-class-template.rst
+
+   lasso_classification
+   lasso_regression
+   plot_lasso_classification
+   plot_lasso_regression
