+++ docs/autosummary/mlcolvar.core.transform.descriptors.CoordinationNumbers.rst
@@ -1,0 +1,40 @@
+﻿mlcolvar.core.transform.descriptors.CoordinationNumbers
+=======================================================
+
+.. currentmodule:: mlcolvar.core.transform.descriptors
+
+.. autoclass:: CoordinationNumbers
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~CoordinationNumbers.__init__
+      ~CoordinationNumbers.compute_coordination_number
+      ~CoordinationNumbers.forward
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~CoordinationNumbers.T_destination
+      ~CoordinationNumbers.call_super_init
+      ~CoordinationNumbers.dump_patches
+      ~CoordinationNumbers.training
+   
+   
+
+
