+++ docs/autosummary/mlcolvar.core.transform.descriptors.EigsAdjMat.rst
@@ -1,0 +1,41 @@
+﻿mlcolvar.core.transform.descriptors.EigsAdjMat
+==============================================
+
+.. currentmodule:: mlcolvar.core.transform.descriptors
+
+.. autoclass:: EigsAdjMat
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~EigsAdjMat.__init__
+      ~EigsAdjMat.compute_adjacency_matrix
+      ~EigsAdjMat.forward
+      ~EigsAdjMat.get_eigenvalues
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~EigsAdjMat.T_destination
+      ~EigsAdjMat.call_super_init
+      ~EigsAdjMat.dump_patches
+      ~EigsAdjMat.training
+   
+   
+
+
