+++ docs/autosummary/mlcolvar.core.transform.descriptors.MultipleDescriptors.rst
@@ -1,0 +1,39 @@
+﻿mlcolvar.core.transform.descriptors.MultipleDescriptors
+=======================================================
+
+.. currentmodule:: mlcolvar.core.transform.descriptors
+
+.. autoclass:: MultipleDescriptors
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~MultipleDescriptors.__init__
+      ~MultipleDescriptors.forward
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~MultipleDescriptors.T_destination
+      ~MultipleDescriptors.call_super_init
+      ~MultipleDescriptors.dump_patches
+      ~MultipleDescriptors.training
+   
+   
+
+
