+++ docs/autosummary/mlcolvar.core.transform.descriptors.PairwiseDistances.rst
@@ -1,0 +1,40 @@
+﻿mlcolvar.core.transform.descriptors.PairwiseDistances
+=====================================================
+
+.. currentmodule:: mlcolvar.core.transform.descriptors
+
+.. autoclass:: PairwiseDistances
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~PairwiseDistances.__init__
+      ~PairwiseDistances.compute_pairwise_distances
+      ~PairwiseDistances.forward
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~PairwiseDistances.T_destination
+      ~PairwiseDistances.call_super_init
+      ~PairwiseDistances.dump_patches
+      ~PairwiseDistances.training
+   
+   
+
+
