+++ docs/autosummary/mlcolvar.core.transform.descriptors.TorsionalAngle.rst
@@ -1,0 +1,41 @@
+﻿mlcolvar.core.transform.descriptors.TorsionalAngle
+==================================================
+
+.. currentmodule:: mlcolvar.core.transform.descriptors
+
+.. autoclass:: TorsionalAngle
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~TorsionalAngle.__init__
+      ~TorsionalAngle.compute_torsional_angle
+      ~TorsionalAngle.forward
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~TorsionalAngle.MODES
+      ~TorsionalAngle.T_destination
+      ~TorsionalAngle.call_super_init
+      ~TorsionalAngle.dump_patches
+      ~TorsionalAngle.training
+   
+   
+
+
