+++ docs/autosummary/mlcolvar.core.transform.tools.ContinuousHistogram.rst
@@ -1,0 +1,40 @@
+﻿mlcolvar.core.transform.tools.ContinuousHistogram
+=================================================
+
+.. currentmodule:: mlcolvar.core.transform.tools
+
+.. autoclass:: ContinuousHistogram
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~ContinuousHistogram.__init__
+      ~ContinuousHistogram.compute_hist
+      ~ContinuousHistogram.forward
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~ContinuousHistogram.T_destination
+      ~ContinuousHistogram.call_super_init
+      ~ContinuousHistogram.dump_patches
+      ~ContinuousHistogram.training
+   
+   
+
+
