+++ docs/autosummary/mlcolvar.core.transform.tools.Normalization.rst
@@ -1,0 +1,44 @@
+﻿mlcolvar.core.transform.tools.Normalization
+===========================================
+
+.. currentmodule:: mlcolvar.core.transform.tools
+
+.. autoclass:: Normalization
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~Normalization.__init__
+      ~Normalization.extra_repr
+      ~Normalization.forward
+      ~Normalization.inverse
+      ~Normalization.set_custom
+      ~Normalization.set_from_stats
+      ~Normalization.setup_from_datamodule
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~Normalization.T_destination
+      ~Normalization.call_super_init
+      ~Normalization.dump_patches
+      ~Normalization.training
+   
+   
+
+
