+++ docs/autosummary/mlcolvar.core.transform.tools.SwitchingFunctions.rst
@@ -1,0 +1,42 @@
+﻿mlcolvar.core.transform.tools.SwitchingFunctions
+================================================
+
+.. currentmodule:: mlcolvar.core.transform.tools
+
+.. autoclass:: SwitchingFunctions
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   .. rubric:: Methods
+
+   .. autosummary::
+   
+      ~SwitchingFunctions.Fermi_switch
+      ~SwitchingFunctions.Rational_switch
+      ~SwitchingFunctions.__init__
+      ~SwitchingFunctions.forward
+   
+   
+
+
+..
+   
+   
+   .. rubric:: Attributes
+
+   .. autosummary::
+   
+      ~SwitchingFunctions.SWITCH_FUNCS
+      ~SwitchingFunctions.T_destination
+      ~SwitchingFunctions.call_super_init
+      ~SwitchingFunctions.dump_patches
+      ~SwitchingFunctions.training
+   
+   
+
+
