+++ docs/autosummary/mlcolvar.cvs.Committor.rst
@@ -1,5 +1,5 @@
 ﻿mlcolvar.cvs.Committor
-====================
+======================
 
 .. currentmodule:: mlcolvar.cvs
