+++ docs/autosummary/mlcolvar.data.DictDataset.rst
@@ -29,6 +29,7 @@
 
    .. autosummary::
    
+      ~DictDataset.feature_names
       ~DictDataset.keys
