+++ docs/autosummary/mlcolvar.explain.lasso.lasso_classification.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.explain.lasso.lasso\_classification
+============================================
+
+.. currentmodule:: mlcolvar.explain.lasso
+
+.. autoclass:: lasso_classification
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
