+++ docs/autosummary/mlcolvar.explain.lasso.lasso_regression.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.explain.lasso.lasso\_regression
+========================================
+
+.. currentmodule:: mlcolvar.explain.lasso
+
+.. autoclass:: lasso_regression
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
