+++ docs/autosummary/mlcolvar.explain.lasso.plot_lasso_classification.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.explain.lasso.plot\_lasso\_classification
+==================================================
+
+.. currentmodule:: mlcolvar.explain.lasso
+
+.. autoclass:: plot_lasso_classification
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
