+++ docs/autosummary/mlcolvar.explain.lasso.plot_lasso_regression.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.explain.lasso.plot\_lasso\_regression
+==============================================
+
+.. currentmodule:: mlcolvar.explain.lasso
+
+.. autoclass:: plot_lasso_regression
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
