+++ docs/autosummary/mlcolvar.explain.sensitivity.plot_sensitivity.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.explain.sensitivity.plot\_sensitivity
+==============================================
+
+.. currentmodule:: mlcolvar.explain.sensitivity
+
+.. autoclass:: plot_sensitivity
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
