+++ docs/autosummary/mlcolvar.explain.sensitivity.sensitivity_analysis.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.explain.sensitivity.sensitivity\_analysis
+==================================================
+
+.. currentmodule:: mlcolvar.explain.sensitivity
+
+.. autoclass:: sensitivity_analysis
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
