+++ docs/autosummary/mlcolvar.utils.plot.plot_features_distribution.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.utils.plot.plot\_features\_distribution
+================================================
+
+.. currentmodule:: mlcolvar.utils.plot
+
+.. autoclass:: plot_features_distribution
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
