+++ docs/autosummary/mlcolvar.utils.plot.plot_metrics.rst
@@ -1,0 +1,23 @@
+﻿mlcolvar.utils.plot.plot\_metrics
+=================================
+
+.. currentmodule:: mlcolvar.utils.plot
+
+.. autoclass:: plot_metrics
+   :members:                                   
+   :show-inheritance:                           
+   :inherited-members: Module,LightningModule                       
+
+   
+   .. automethod:: __init__
+
+   
+   
+
+
+..
+   
+   
+   
+
+
