+++ docs/examples_experiments.rst
@@ -1,7 +1,7 @@
 Modified Muller-Brown
 =====================
 
-Experiments on the modified Muller-Brown potential (3 states) reported in the mlcolvar' [JCP paper](https://doi.org/10.1063/5.0156343), describing different learning scenarios. 
+Experiments on the modified Muller-Brown potential (3 states) reported in the mlcolvar' JCP paper (https://doi.org/10.1063/5.0156343), describing different learning scenarios. 
 
 .. toctree::
    :maxdepth: 1
