+++ docs/tutorials_advanced.rst
@@ -8,4 +8,3 @@ Customizing CVs
    notebooks/tutorials/adv_newcv_scratch.ipynb
    notebooks/tutorials/adv_newcv_subclass.ipynb
    notebooks/tutorials/adv_preprocessing.ipynb
-   notebooks/tutorials/adv_features_relevances.ipynb
