+++ mlcolvar/explain/__init__.py
@@ -1,9 +1,7 @@
 __all__ = [
     "sensitivity_analysis", 
     "plot_sensitivity",
-    "plot_features_distribution"
 ]
  
 from .sensitivity import *
-from .utils import *
 # from .lasso import * # lasso requires additional dependencies
